Add CRUD для Users

+++  BL/Users/Entities/CreateUserModel.cs
@@ -9,7 +9,7 @@ public class CreateUserModel
     public string Email { get; set; }
     public string PasswordHash { get; set; }
     
-    public int RoleId { get; set; }
-    public int DriverLicenseId { get; set; }
-    public int StateId { get; set; }
+    public long RoleId { get; set; }
+    public long DriverLicenseId { get; set; }
+    public long StateId { get; set; }
 }
