+++  BL/Users/Provider/IUserProvider.cs
@@ -5,5 +5,5 @@ namespace BL.Users.Provider;
 public interface IUserProvider
 {
     IEnumerable<UserModel> GetUsers(FilterUserModel filter = null);
-    UserModel GerUserInfo(int id);
+    UserModel GerUserInfo(long id);
 }
