+++ DataAccess/Entities/RentEntity.cs
@@ -9,12 +9,12 @@ public class RentEntity : BaseEntity
     public DateTime EndDate { get; set; }
     public decimal Cost { get; set; }
 
-    public int CarId { get; set; }
+    public long CarId { get; set; }
     public virtual CarEntity Car { get; set; }
 
-    public int UserId { get; set; }
+    public long UserId { get; set; }
     public virtual UserEntity User { get; set; }
 
-    public int StateId { get; set; }
+    public long StateId { get; set; }
     public virtual StateEntity State { get; set; }
 }
