+++ DataAccess/Entities/TechnicalInspectionEntity.cs
@@ -9,12 +9,12 @@ public class TechnicalInspectionEntity : BaseEntity
     public DateTime InspectionEnd { get; set; }
     public string InspectionResult { get; set; }
 
-    public int MaintenanceStateId { get; set; }
+    public long MaintenanceStateId { get; set; }
     public virtual StateEntity State { get; set; }
 
-    public int TechnicianId { get; set; }
+    public long TechnicianId { get; set; }
     public virtual UserEntity TechnicalEmployee { get; set; }
 
-    public int CarId { get; set; }
+    public long CarId { get; set; }
     public virtual CarEntity Car { get; set; }
 }
